+++ prototype/analyzer/deprecated/c_builtins/libc.py
