+++ prototype/analyzer/deprecated/lol_analyzer_reserved_names.py
