+++ prototype/examples/helloworld.lol
@@ -1,7 +1,7 @@
 ### Basic hello world
 module io = import("stdio.h");
 
-function main() -> int32 {
-    printf("Hello, World!\n");
+function main() -> i32 {
+    io::printf("Hello, World!\n");
     return 0;
 }
