+++ prototype/parser/lol_parser_token_stream.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from lexer.lol_lexer import Token
+from prototype.lexer.lol_lexer import Token
 
 
 class TokenStream:
